buoi 4

+++ app/Http/Controllers/Admin/CatalogueController.php
@@ -17,7 +17,7 @@ class CatalogueController extends Controller
      */
     public function index()
     {
-        $data=Catelogue::query()->latest('id')->paginate(5);
+        $data=Catelogue::query()->latest('id')->get();
         return view(self::PATH_VIEW.__FUNCTION__,compact('data'));
     }
 
@@ -80,7 +80,7 @@ class CatalogueController extends Controller
         $currenCover=$model->cover;
 
         $model->update($data);
-        if($currenCover && Storage::exists($currenCover)){
+        if($request->hasFile('cover') && $currenCover && Storage::exists($currenCover)){
             Storage::delete($currenCover);
         }
         return back();
